chore(interfaces): move interfaces to separate files
