+++ frontend/src/interfaces/clientState.ts
@@ -1,0 +1,7 @@
+import type { Product } from "./product"
+
+export interface ClientState {
+    isBurgerMenuOpen: boolean
+    currentPage: string | null
+    products: Product[]
+}
