+++ frontend/src/interfaces/introBlockProps.ts
@@ -1,0 +1,4 @@
+export interface IntroBlockProps {
+    content: React.ReactNode,
+    backgroundImageURL: string
+}
