+++ frontend/src/interfaces/newsWrapProps.ts
@@ -1,0 +1,3 @@
+export interface NewsWrapProps {
+    defaultNewsCount: number
+}
