+++ frontend/src/interfaces/portfolioItem.ts
@@ -1,0 +1,11 @@
+export interface PortfolioItem {
+    id: number,
+    title: string
+    description: string
+    date: string
+    client: string
+    category: string
+    service: string
+    type: string
+    imageURL: string
+}
