+++ frontend/src/interfaces/portfolioItemCardProps.ts
@@ -1,0 +1,7 @@
+export interface PortfolioItemCardProps {
+    id: number,
+    title: string,
+    type: string,
+    backgroundImageURL: string,
+    className: string
+}
