+++ frontend/src/interfaces/product.ts
@@ -1,0 +1,10 @@
+export interface Product {
+    id: number
+    title: string
+    type: string
+    imageURL: string
+    rating: number
+    price: number
+    discountPrice: number
+    description: string
+}
