+++ frontend/src/interfaces/productCardProps.ts
@@ -1,0 +1,3 @@
+export interface ProductCardProps {
+    isCartItem?: boolean
+}
