+++ frontend/src/interfaces/review.ts
@@ -1,0 +1,7 @@
+export interface Review {
+    avatarURL: string,
+    rating: number,
+    text: string,
+    name: string,
+    post: string
+}
