+++ frontend/src/interfaces/subIntroBlockProps.ts
@@ -1,0 +1,4 @@
+export interface SubIntroBlockProps {
+    title: string,
+    backgroundImageURL: string
+}
