+++ frontend/src/interfaces/teamExpert.ts
@@ -1,0 +1,10 @@
+export interface TeamExpert {
+    imageURL: string
+    name: string
+    jobTitle: string
+    socialNetworksURLs: {
+        instagramURL?: string
+        facebookURL: string
+        twitterURL?: string
+    }
+}
