+++ frontend/src/interfaces/teamSectionProps.ts
@@ -1,0 +1,3 @@
+export interface TeamSectionProps {
+    defaultExpertsCount: number
+}
