+++ frontend/src/interfaces/utilsIntroBlockProps.ts
@@ -1,0 +1,3 @@
+export interface UtilsIntroBlockProps {
+    title: string
+}
