+++ frontend/src/store/index.ts
@@ -10,4 +10,4 @@ const store = configureStore({
 export type RootState = ReturnType<typeof store.getState>
 export type AppDispatch = typeof store.dispatch
 
-export default store 
+export default store
